preload: restore shared api.ts and disable sandbox on host chrome views
Reverts the temporary inline-duplication workaround from e074738 in
favor of Option 1: the title-bar and panel WebContentsView host
chrome surfaces now run with sandbox: false, which lets the preload
entries require() the shared out/preload/chunks/api-*.js that Rollup
emits from src/preload/api.ts.

Why:
- Electron 40 defaults preloads to sandbox: true.
- Sandboxed preloads can only require() from electron/events/timers/url.
- The shared chunk require silently failed under the sandbox, leaving
  window.api undefined and blanking the renderer (and breaking
  renderer-side Datadog/PostHog telemetry as a side effect).
- Inlining the API into every preload kept dev working but duplicated
  ~330 lines and made every bridge change a two-place edit.

Approach:
- src/preload/api.ts is the single source of truth for buildElectronApi().
- src/preload/index.ts and src/preload/comfyTitleBarPreload.ts both
  import buildElectronApi() from ./api.
- src/main/index.ts sets sandbox: false on the title-bar
  WebContentsView and the panel WebContentsView only.
- comfyPreload.js (body view) and comfyTitleMenuPreload.js are
  self-contained and keep their default sandbox posture.
- contextIsolation: true and nodeIntegration: false remain enabled
  on every surface, so the real renderer/Node boundary is unchanged.

Future work tracked in issue #521: build-time chunk inlining plugin
that will let us re-enable sandbox without re-introducing the source
duplication.

Amp-Thread-ID: https://ampcode.com/threads/T-019e0761-2539-7519-af25-e7263fa60eed
